+++ docs/versionhistory.rst
@@ -7,7 +7,7 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
 - Added an asynchronous implementation of the ``itertools`` module
   (`#998 <https://github.com/agronholm/anyio/issues/998>`_; PR by @11kkw)
-- Added the ``local_port`` parameter to :func:`connect_tcp` to allow binding to a
+- Added the ``local_port`` parameter to ``connect_tcp()`` to allow binding to a
   specific local port before connecting
   (`#1067 <https://github.com/agronholm/anyio/issues/1067>`_; PR by @nullwiz)
 - Added support for custom capacity limiters in async path and file I/O
@@ -18,6 +18,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - Changed ``TaskGroup.start_soon()`` to return a ``TaskHandle``
 - Added an option for ``TaskGroup.start()`` to return a ``TaskHandle`` (which then
   contains the start value in the ``start_value`` property)
+- Added the ``cancel()`` convenience method to ``TaskGroup`` as a shortcut for
+  cancelling the task group's cancel scope
 - Improved the error message when a known backend is not installed to suggest the
   install command
   (`#1115 <https://github.com/agronholm/anyio/pull/1115>`_; PR by @EmmanuelNiyonshuti)